requirements.txt
openai
llama_index
pdfplumber
python-dotenv
streamlit